requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
autopep8==1.6.0
boto3==1.20.43
botocore==1.23.43
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==3.0.1
cloudinary==1.32.0
colorama==0.4.4
constantly==15.1.0
crispy-bootstrap5==0.6
cryptography==36.0.1
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-crispy-forms==1.13.0
django-extensions==3.1.5
django-filter==21.1
django-model-utils==4.2.0
django-on-heroku==1.1.2
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gunicorn==20.1.0
hiredis==2.0.0
humanize==3.13.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.0
incremental==21.3.0
iniconfig==1.1.1
install==1.3.5
jmespath==0.10.0
Markdown==3.3.6
msgpack==1.0.3
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
protobuf==4.22.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pymemcache==3.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
requests==2.28.2
rsa==4.9
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.4.0
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.4.0
zipp==3.6.0
zope.interface==5.4.0
